Handling WSL/MSYS unicode path escapes
During conversion of path names from UTF8 to UTF16, characters that
are allowed according to POSIX but not Win32 (i.e. "*:<>?|) are escaped
according to the same rules that WSL and MSYS use. In most conversions
from UTF16 to UTF8, equivalent unescaping is performed, so that only the
plain POSIX representation of the characters is used within Git.

This enables easier use of repos that have file names with characters
that are disallowed according to Win32 path rules, save for backslash.
Because there is currently no way to escape filenames in the git output
(e.g. ls-files), other Win32 tools that parse git output must also be
aware of the escaping rules to work properly. This is true for the
MSYS tools, but not e.g. VSCode. Note that this limitation isn't new, but
represents further work that can be done in this area.